requirements.txt
python-telegram-bot==22.2
asyncio==3.4.3
aiofiles==24.1.0

